+++ src/app/ClientEntryPoint.tsx
@@ -21,48 +21,19 @@ export default function ClientEntryPoint({
 }) {
   return (
     <>
-      <AppShell
-        header={{ height: 100 }}
-        navbar={{
-          width: "25%",
-          breakpoint: "xs",
-          collapsed: { mobile: false, desktop: false },
-        }}
-        padding="md"
-      >
-        <AppShell.Header className={pageStyles["app-header"]}>
-          <Group h="100%" px="md">
-            <IconWriting
-              style={{ width: rem(50), height: rem(50) }}
-              stroke={2}
-            ></IconWriting>
+      <AppShell.Navbar p="md">
+        <AppShell.Section>
+          <SearchBar />
+        </AppShell.Section>
 
-            <div>
-              <Text size="lg" fw="500">
-                A simple notes application by Vinny DeFrancesco
-              </Text>
+        <AppShell.Section grow my="md" component={ScrollArea}>
+          <NotesList notes={notes} />
+        </AppShell.Section>
+      </AppShell.Navbar>
 
-              <Text size="sm">
-                Do you like writing notes? You&apos;re in luck!
-              </Text>
-            </div>
-          </Group>
-        </AppShell.Header>
-
-        <AppShell.Navbar p="md">
-          <AppShell.Section>
-            <SearchBar />
-          </AppShell.Section>
-
-          <AppShell.Section grow my="md" component={ScrollArea}>
-            <NotesList notes={notes} />
-          </AppShell.Section>
-        </AppShell.Navbar>
-
-        <AppShell.Main>
-          <SelectedNote note={note} patients={patients} />
-        </AppShell.Main>
-      </AppShell>
+      <AppShell.Main>
+        <SelectedNote note={note} patients={patients} />
+      </AppShell.Main>
     </>
   );
 }
