+++ src/app/notes_application/patients/page.tsx
@@ -1,0 +1,8 @@
+import { getPatients } from "@/app/server/serverActions";
+import PatientsClientEntryPoint from "@/app/notes_application/patients/PatientsClientEntryPoint";
+
+export default async function PatientsPage() {
+  const patients = await getPatients();
+
+  return <PatientsClientEntryPoint patients={patients} />;
+}
